+++ src/components/dialogs/BottomSheet.tsx
@@ -22,6 +22,8 @@ export type BottomSheetProps = {
   titleStyle?: StyleProp<ViewStyle>;
   /** Close callback */
   onClose?: () => void;
+  /** Hide callback (triggered after close animation finishes) */
+  onHide?: () => void;
   /** Open callback */
   onOpen?: () => void;
 };
@@ -51,6 +53,7 @@ const BottomSheet = forwardRef<BottomSheetRef, BottomSheetProps>((props: BottomS
     titleRight,
     titleStyle = {},
     onClose,
+    onHide: onHideProp,
     onOpen,
   } = props;
 
@@ -112,11 +115,13 @@ const BottomSheet = forwardRef<BottomSheetRef, BottomSheetProps>((props: BottomS
     onOpen?.();
   };
 
+  /** Modal hide callback (after animation finishes) */
   const onHide = () => {
     // Call the closure callback once hide animation has finished, at which point another modal
     //   could be displayed (earliest point). However, this still isn't quite enough and sometimes
     //   requires waiting an additional brief period to ensure modal has completely hidden.
     setTimeout(() => {
+      onHideProp?.();
       onHideCallbackRef.current?.();
     }, 10);
 
@@ -147,7 +152,11 @@ const BottomSheet = forwardRef<BottomSheetRef, BottomSheetProps>((props: BottomS
       >
         {Boolean(title) && (
           <View style={[styles.sheetTitle, inset ? undefined : styles.sheetInset]}>
-            <Text style={[styles.sheetTitleText, titleStyle]} variant="titleLarge">
+            <Text
+              numberOfLines={1}
+              style={[styles.sheetTitleText, titleStyle]}
+              variant="titleLarge"
+            >
               {title}
             </Text>
             {titleRight}
@@ -181,6 +190,8 @@ const styles = StyleSheet.create({
     marginBottom: BOTTOM_SHEET_PADDING,
   },
   sheetTitleText: {
+    flexGrow: 0,
+    flexShrink: 1,
     marginRight: "auto",
   },
 });