+++ src/components/icons/index.ts
@@ -1,1 +1,2 @@
+export { default as DayIcon } from "./DayIcon";
 export { default as LanguageIcon } from "./LanguageIcon";