+++ CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.3.0] - 2023-02-10
+
 ### Added
 
+- View date details and various time units
+- Change individual date display time units
 - Store URL in About screen developer chips
+- Behaviour setting to swap theme on "Details" screen
+
+### Updated
+
+- Abstracted common date actions into hooks
+
 
 ## [0.2.0] - 2023-02-09
 
@@ -17,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to share days via config link
 - Behaviour setting to confirm adding a shared day
-- Show build Git hash and application ID on Debug page
+- Show build Git hash and application ID on "Debug" screen
 - Ability to copy device debug info
 
 ### Changed
@@ -25,16 +36,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support multiple installs with dynamic application ID
 - Indicate when countdown/countup day is current date
 
+
 ## [0.1.1] - 2023-02-05
 
 ### Changed
 
-- Update Debug screen values for EAS Updates*
+- Update "Debug" screen values for EAS Updates*
 
 ### Fixed
 
 - Prevent date selection dialog flicker when transitioning workflows
 - Lighten initial splash screen color
+
 
 ## [0.1.0] - 2023-02-04